requirements.txt
google-generativeai==0.7.2
openpyxl==3.1.2
numpy==1.23.2
pandas==1.4.3